Add tests for Event
