+++ src/sink/udp.rs
@@ -79,10 +79,10 @@ impl Sink for Udp {
         if success {
             Ok(Some(vec![sink::Reply::Insight(event.insight_ack())]))
         } else {
-            Ok(event
-                .insight_trigger()
-                .and_then(|e1| event.insight_fail().map(|e2| (e1, e2)))
-                .map(|(e1, e2)| vec![sink::Reply::Insight(e1), sink::Reply::Insight(e2)]))
+            Ok(Some(vec![
+                sink::Reply::Insight(event.insight_trigger()),
+                sink::Reply::Insight(event.insight_fail()),
+            ]))
         }
     }
     fn default_codec(&self) -> &str {
